+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
+- 2018-02-09 : (PLAM-2106)  : Added functionality to scan for peripherals in background
+
 ### Changed
 ### Fixed
 - 2018-02-09 : (PLAM-2282)  : Warnings in dependencies resolved